Move crack-related methods from Enigma to Revert module
