+++ test/test_helper.rb
@@ -9,3 +9,4 @@ require './lib/key'
 require './lib/offset'
 require './lib/shift'
 require './lib/enigma'
+require './modules/revert'
